test(elementor-v4): split content-walker tests into focused files
Replaces the single content-walker.test.js with three focused files:
prop-unwrappers.test.js, extractors.test.js, walk-atomic-tree.test.js.
Shared node builders extracted to __fixtures__/nodes.js. Merges the
two redundant text-editor extractor tests (external vs internal anchor)
into one test covering both.
