+++ packages/js/jest.config.js
@@ -9,6 +9,7 @@ module.exports = {
 		"/tests/setupTests.js",
 		"/tests/decorator/__mocks__/@wordpress/rich-text/index.js",
 		"/tests/test-utils.js",
+		"/tests/elementor-v4/__fixtures__/",
 	],
 	// https://testing-library.com/docs/react-testing-library/setup#jest-27
 	testEnvironment: "jest-environment-jsdom",
